Settings consolidation U2-U5: extract review tabs; global dialog is monolith-free
Sequential extraction (workflow waremxcbr, per-unit adversarial verify):
- U2 GitTab -> settings/ReviewGitTab.tsx
- U3 ReviewDisplayTab -> settings/ReviewDisplayTab.tsx (co-located SegmentedControl/ToggleSwitch for visual parity)
- U4 CommentsTab (+ CCLabel helpers) -> settings/CommentsTab.tsx
- U5 AppSettingsDialog repointed to the new files -> imports NOTHING from the monolith

All config-store-only, byte-for-byte parity verified. Monolith still compiles (its own
JSX + the portal/standalone path) until later units. Verified: frontend typecheck +
build, AND build:portal green.
