+++ packages/ui/components/Settings.tsx
@@ -3,7 +3,6 @@ import { createPortal } from 'react-dom';
 import type { Origin } from '@plannotator/shared/agents';
 import type { DiffLineBgIntensity } from '@plannotator/shared/config';
 import { configStore, useConfigValue } from '../config';
-import { loadDiffFont } from '../utils/diffFonts';
 import { TaterSpritePullup } from './TaterSpritePullup';
 import { getIdentity, regenerateIdentity, setCustomIdentity } from '../utils/identity';
 import { GitUser } from '../icons/GitUser';
@@ -40,6 +39,8 @@ import { isMac, altKey } from '../utils/platform';
 import { getAIProviderSettings, resolveAIProviderSelection } from '../utils/aiProvider';
 import { AISettingsTab } from './AISettingsTab';
 import { HooksTab } from './settings/HooksTab';
+import { GitTab } from './settings/ReviewGitTab';
+import { CommentsTab } from './settings/CommentsTab';
 import { OverlayScrollArea } from './OverlayScrollArea';
 import {
   getFileBrowserSettings,
@@ -66,40 +67,8 @@ interface SettingsProps {
 }
 
 // --- Review-mode Display tab (diff display options) ---
-// Constants now live in ./settings/diffOptions (this monolith is being deleted).
-import {
-  DIFF_FONT_OPTIONS,
-  DIFF_STYLE_OPTIONS,
-  OVERFLOW_OPTIONS,
-  INDICATOR_OPTIONS,
-  LINE_DIFF_OPTIONS,
-  LINE_BG_INTENSITY_OPTIONS,
-  DEFAULT_DIFF_TYPE_OPTIONS,
-} from './settings/diffOptions';
-
-function SegmentedControl<T extends string>({ options, value, onChange }: {
-  options: { value: T; label: string }[];
-  value: T;
-  onChange: (v: T) => void;
-}) {
-  return (
-    <div className="inline-flex items-center gap-0.5 bg-muted/50 rounded-lg p-0.5">
-      {options.map((opt) => (
-        <button
-          key={opt.value}
-          onClick={() => onChange(opt.value)}
-          className={`px-2.5 py-1 text-xs rounded-md whitespace-nowrap transition-colors ${
-            value === opt.value
-              ? 'bg-background text-foreground shadow-sm font-medium'
-              : 'text-muted-foreground hover:text-foreground'
-          }`}
-        >
-          {opt.label}
-        </button>
-      ))}
-    </div>
-  );
-}
+// Extracted to ./settings/ReviewDisplayTab (this monolith is being deleted).
+import { ReviewDisplayTab } from './settings/ReviewDisplayTab';
 
 function ToggleSwitch({ checked, onChange, label, description }: {
   checked: boolean;
@@ -130,393 +99,6 @@ function ToggleSwitch({ checked, onChange, label, description }: {
     </div>
   );
 }
-
-export const GitTab: React.FC = () => {
-  const defaultDiffType = useConfigValue('defaultDiffType');
-  return (
-    <div className="space-y-2">
-      <div>
-        <div className="text-sm font-medium">Default Diff View</div>
-        <div className="text-xs text-muted-foreground">Which changes to show when you open a code review</div>
-      </div>
-      <div className="space-y-2">
-        {DEFAULT_DIFF_TYPE_OPTIONS.map((opt) => (
-          <button
-            key={opt.value}
-            type="button"
-            onClick={() => configStore.getState().set('defaultDiffType', opt.value)}
-            className={`w-full flex items-start gap-3 p-3 rounded-lg border transition-colors text-left ${
-              defaultDiffType === opt.value
-                ? 'border-primary bg-primary/5'
-                : 'border-border hover:border-muted-foreground/30 hover:bg-muted/50'
-            }`}
-          >
-            <div className={`mt-0.5 w-4 h-4 rounded-full border-2 flex-shrink-0 flex items-center justify-center ${
-              defaultDiffType === opt.value ? 'border-primary' : 'border-muted-foreground/40'
-            }`}>
-              {defaultDiffType === opt.value && (
-                <div className="w-2 h-2 rounded-full bg-primary" />
-              )}
-            </div>
-            <div>
-              <div className="text-sm font-medium">{opt.label}</div>
-              <div className="text-xs text-muted-foreground">{opt.description}</div>
-            </div>
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-export const ReviewDisplayTab: React.FC = () => {
-  const diffStyle = useConfigValue('diffStyle');
-  const diffOverflow = useConfigValue('diffOverflow');
-  const diffIndicators = useConfigValue('diffIndicators');
-  const diffLineDiffType = useConfigValue('diffLineDiffType');
-  const diffShowLineNumbers = useConfigValue('diffShowLineNumbers');
-  const diffShowBackground = useConfigValue('diffShowBackground');
-  const diffLineBgIntensity = useConfigValue('diffLineBgIntensity');
-  const diffHideWhitespace = useConfigValue('diffHideWhitespace');
-  const diffFontFamily = useConfigValue('diffFontFamily');
-  const diffFontSize = useConfigValue('diffFontSize');
-  const diffTabSize = useConfigValue('diffTabSize');
-
-  useEffect(() => {
-    if (diffFontFamily) loadDiffFont(diffFontFamily);
-  }, [diffFontFamily]);
-
-  const fontSizeNum = diffFontSize ? parseInt(diffFontSize) : 13;
-
-  return (
-    <div className="space-y-6">
-      {/* Typography */}
-      <section className="space-y-3">
-        <h3 className="text-[10px] font-medium uppercase tracking-wider text-muted-foreground">Typography</h3>
-        <div className="space-y-2">
-          <div className="flex items-center justify-between">
-            <div className="text-sm font-medium">Code Font</div>
-            <select
-              value={diffFontFamily}
-              onChange={(e) => configStore.getState().set('diffFontFamily', e.target.value)}
-              className="px-2 py-1 text-xs rounded-md bg-muted/50 border border-border text-foreground"
-              style={diffFontFamily ? { fontFamily: `'${diffFontFamily}', monospace` } : undefined}
-            >
-              {DIFF_FONT_OPTIONS.map((opt) => (
-                <option key={opt.value} value={opt.value}>{opt.label}</option>
-              ))}
-            </select>
-          </div>
-          <div className="flex items-center justify-between">
-            <div className="text-sm font-medium">Font Size</div>
-            <div className="flex items-center gap-1">
-              <button
-                type="button"
-                onClick={() => configStore.getState().set('diffFontSize', `${Math.max(8, fontSizeNum - 1)}px`)}
-                disabled={fontSizeNum <= 8}
-                className="h-7 w-7 rounded-md bg-muted text-foreground flex items-center justify-center text-xs font-medium disabled:opacity-30"
-              >
-                −
-              </button>
-              <span className="w-8 text-center text-xs font-mono tabular-nums">{fontSizeNum}px</span>
-              <button
-                type="button"
-                onClick={() => configStore.getState().set('diffFontSize', `${Math.min(24, fontSizeNum + 1)}px`)}
-                disabled={fontSizeNum >= 24}
-                className="h-7 w-7 rounded-md bg-muted text-foreground flex items-center justify-center text-xs font-medium disabled:opacity-30"
-              >
-                +
-              </button>
-            </div>
-          </div>
-          <div className="flex items-center justify-between">
-            <div className="text-sm font-medium">Tab Size</div>
-            <div className="flex items-center gap-1">
-              <button
-                type="button"
-                onClick={() => configStore.getState().set('diffTabSize', Math.max(1, diffTabSize - 1))}
-                disabled={diffTabSize <= 1}
-                className="h-7 w-7 rounded-md bg-muted text-foreground flex items-center justify-center text-xs font-medium disabled:opacity-30"
-              >
-                −
-              </button>
-              <span className="w-8 text-center text-xs font-mono tabular-nums">{diffTabSize}</span>
-              <button
-                type="button"
-                onClick={() => configStore.getState().set('diffTabSize', Math.min(8, diffTabSize + 1))}
-                disabled={diffTabSize >= 8}
-                className="h-7 w-7 rounded-md bg-muted text-foreground flex items-center justify-center text-xs font-medium disabled:opacity-30"
-              >
-                +
-              </button>
-            </div>
-          </div>
-        </div>
-        <div
-          className="text-xs text-muted-foreground px-2 py-1.5 rounded-md bg-muted/30 font-mono"
-          style={{ fontFamily: diffFontFamily ? `'${diffFontFamily}', monospace` : undefined, fontSize: `${fontSizeNum}px` }}
-        >
-          const x = fn(42);
-        </div>
-      </section>
-
-      <div className="border-t border-border" />
-
-      {/* Diff Layout */}
-      <section className="space-y-3">
-        <h3 className="text-[10px] font-medium uppercase tracking-wider text-muted-foreground">Layout</h3>
-        <div className="flex items-center justify-between gap-4">
-          <div className="text-sm font-medium shrink-0">Diff Style</div>
-          <SegmentedControl options={DIFF_STYLE_OPTIONS} value={diffStyle} onChange={(v) => configStore.getState().set('diffStyle', v)} />
-        </div>
-        <div className="flex items-center justify-between gap-4">
-          <div className="text-sm font-medium shrink-0">Line Overflow</div>
-          <SegmentedControl options={OVERFLOW_OPTIONS} value={diffOverflow} onChange={(v) => configStore.getState().set('diffOverflow', v)} />
-        </div>
-        <div className="flex items-center justify-between gap-4">
-          <div className="text-sm font-medium shrink-0">Change Indicators</div>
-          <SegmentedControl options={INDICATOR_OPTIONS} value={diffIndicators} onChange={(v) => configStore.getState().set('diffIndicators', v)} />
-        </div>
-        <div className="flex items-center justify-between gap-4">
-          <div className="text-sm font-medium shrink-0">Inline Diff</div>
-          <SegmentedControl options={LINE_DIFF_OPTIONS} value={diffLineDiffType} onChange={(v) => configStore.getState().set('diffLineDiffType', v)} />
-        </div>
-      </section>
-
-      <div className="border-t border-border" />
-
-      {/* Toggles */}
-      <section className="space-y-3">
-        <h3 className="text-[10px] font-medium uppercase tracking-wider text-muted-foreground">Options</h3>
-        <ToggleSwitch
-          checked={diffShowLineNumbers}
-          onChange={(v) => configStore.getState().set('diffShowLineNumbers', v)}
-          label="Line Numbers"
-        />
-        <ToggleSwitch
-          checked={diffHideWhitespace}
-          onChange={(v) => configStore.getState().set('diffHideWhitespace', v)}
-          label="Hide Whitespace"
-        />
-        <ToggleSwitch
-          checked={diffShowBackground}
-          onChange={(v) => configStore.getState().set('diffShowBackground', v)}
-          label="Line Backgrounds"
-        />
-        {diffShowBackground && (
-          <div className="flex items-center justify-between gap-4 pl-4">
-            <div className="text-sm font-medium shrink-0">Intensity</div>
-            <SegmentedControl options={LINE_BG_INTENSITY_OPTIONS} value={diffLineBgIntensity} onChange={(v) => configStore.getState().set('diffLineBgIntensity', v)} />
-          </div>
-        )}
-      </section>
-    </div>
-  );
-};
-
-/**
- * Stored label config — serialized to JSON in the config store.
- * `blocking` controls whether the blocking/non-blocking toggle
- * appears in the picker when this label is selected.
- */
-interface CCLabelConfig {
-  label: string;
-  display: string;
-  blocking: boolean; // true = show toggle in picker, false = no decoration
-}
-
-const DEFAULT_CC_LABELS: CCLabelConfig[] = [
-  { label: 'suggestion', display: 'suggestion', blocking: true },
-  { label: 'nitpick',    display: 'nit',        blocking: false },
-  { label: 'question',   display: 'question',   blocking: true },
-  { label: 'issue',      display: 'issue',      blocking: true },
-  { label: 'praise',     display: 'praise',     blocking: false },
-  { label: 'thought',    display: 'thought',    blocking: false },
-  { label: 'note',       display: 'note',       blocking: false },
-  { label: 'todo',       display: 'todo',       blocking: true },
-  { label: 'chore',      display: 'chore',      blocking: true },
-];
-
-function parseCCLabels(json: string | null): CCLabelConfig[] {
-  if (!json) return DEFAULT_CC_LABELS;
-  try {
-    const parsed = JSON.parse(json);
-    if (!Array.isArray(parsed)) return DEFAULT_CC_LABELS;
-    return parsed.map((l: Record<string, unknown>) => ({
-      label: (l.label as string) || 'custom',
-      display: (l.display as string) || (l.label as string) || 'custom',
-      blocking: l.blocking === true || l.blocking === 'true',
-    }));
-  } catch {
-    return DEFAULT_CC_LABELS;
-  }
-}
-
-export const CommentsTab: React.FC = () => {
-  const conventionalComments = useConfigValue('conventionalComments');
-  const labelsJson = useConfigValue('conventionalLabels');
-  const [labels, setLabels] = useState(() => parseCCLabels(labelsJson));
-
-  const isDefault = labelsJson === null;
-
-  const save = (next: CCLabelConfig[]) => {
-    setLabels(next);
-    configStore.getState().set('conventionalLabels', JSON.stringify(next));
-  };
-
-  const updateLabel = (index: number, updates: Partial<CCLabelConfig>) => {
-    const next = [...labels];
-    next[index] = { ...next[index], ...updates };
-    save(next);
-  };
-
-  const removeLabel = (index: number) => {
-    save(labels.filter((_, i) => i !== index));
-  };
-
-  const addLabel = () => {
-    const existing = new Set(labels.map(l => l.label));
-    let slug = 'custom';
-    let n = 2;
-    while (existing.has(slug)) { slug = `custom-${n++}`; }
-    save([...labels, { label: slug, display: slug, blocking: false }]);
-  };
-
-  const resetToDefaults = () => {
-    setLabels(DEFAULT_CC_LABELS);
-    configStore.getState().set('conventionalLabels', null);
-  };
-
-  return (
-    <>
-      <ToggleSwitch
-        checked={conventionalComments}
-        onChange={(v) => configStore.getState().set('conventionalComments', v)}
-        label="Conventional Comments"
-        description="Add structured labels to review comments"
-      />
-
-      <div className={`space-y-4 transition-opacity ${conventionalComments ? '' : 'opacity-40 pointer-events-none'}`}>
-
-      <div className="border-t border-border" />
-
-      {/* How it works */}
-      <div className="space-y-3">
-        <div>
-          <div className="text-sm font-medium">How it works</div>
-          <div className="text-xs text-muted-foreground leading-relaxed mt-1">
-            When enabled, a label picker appears above the comment input when annotating code.
-            Labels classify your feedback intent, making it clear whether a comment is a blocking issue,
-            a trivial nitpick, or praise.
-          </div>
-        </div>
-
-        <div className="text-xs text-muted-foreground leading-relaxed">
-          Based on the{' '}
-          <a
-            href="https://conventionalcomments.org"
-            target="_blank"
-            rel="noopener noreferrer"
-            className="text-primary hover:underline"
-          >
-            Conventional Comments
-          </a>
-          {' '}spec. Comments are exported as plain text, readable on GitHub, and parseable by tooling.
-        </div>
-
-        {/* Example output */}
-        <div className="rounded-md border border-border bg-muted/30 p-3 space-y-1.5">
-          <div className="text-[10px] font-medium uppercase tracking-wider text-muted-foreground">Example output</div>
-          <div className="font-mono text-[11px] text-foreground/80 leading-relaxed">
-            <span className="font-bold">issue</span> <span className="text-muted-foreground">(blocking)</span>: This will throw if user is null — the guard was removed in the refactor.
-          </div>
-        </div>
-      </div>
-
-      <div className="border-t border-border" />
-
-      {/* Label editor */}
-      <div className="flex items-center justify-between">
-        <div>
-          <div className="text-sm font-medium">Labels</div>
-          <div className="text-xs text-muted-foreground">
-            Customize labels and their default severity
-          </div>
-        </div>
-        {!isDefault && (
-          <button
-            onClick={resetToDefaults}
-            className="text-[10px] text-muted-foreground hover:text-foreground transition-colors"
-          >
-            Reset to defaults
-          </button>
-        )}
-      </div>
-
-      {/* Column headers */}
-      <div className="flex items-end gap-2 px-2 text-[10px] font-medium uppercase tracking-wider text-muted-foreground/60">
-        <span className="flex-1">Label</span>
-        <span className="w-20 text-center">Blocking decorator</span>
-        <span className="w-6" />
-      </div>
-
-      <div className="space-y-1.5">
-        {labels.map((label, index) => (
-          <div
-            key={index}
-            className="flex items-center gap-2 p-2 rounded-lg bg-muted/40"
-          >
-            <input
-              type="text"
-              value={label.display}
-              onChange={(e) => {
-                const display = e.target.value;
-                updateLabel(index, { display, label: display });
-              }}
-              className="flex-1 px-2 py-1 bg-background/80 rounded text-xs font-mono focus:outline-none focus:ring-1 focus:ring-primary/50 min-w-0"
-            />
-            <div className="w-20 flex justify-center">
-              <button
-                role="switch"
-                aria-checked={label.blocking}
-                onClick={() => updateLabel(index, { blocking: !label.blocking })}
-                className={`relative inline-flex h-5 w-9 items-center rounded-full transition-colors ${
-                  label.blocking ? 'bg-primary' : 'bg-muted'
-                }`}
-              >
-                <span
-                  className={`inline-block h-3.5 w-3.5 transform rounded-full bg-white shadow-sm transition-transform ${
-                    label.blocking ? 'translate-x-4.5' : 'translate-x-0.5'
-                  }`}
-                />
-              </button>
-            </div>
-            <button
-              onClick={() => removeLabel(index)}
-              className="p-1 rounded hover:bg-destructive/10 text-muted-foreground hover:text-destructive transition-colors flex-shrink-0"
-              title="Remove label"
-            >
-              <svg className="w-3 h-3" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-                <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-              </svg>
-            </button>
-          </div>
-        ))}
-      </div>
-
-      {labels.length < 12 && (
-        <button
-          onClick={addLabel}
-          className="w-full py-1.5 text-xs text-muted-foreground hover:text-foreground border border-dashed border-border rounded-lg hover:border-foreground/30 transition-colors"
-        >
-          + Add label
-        </button>
-      )}
-
-      </div>
-    </>
-  );
-};
 
 export const Settings: React.FC<SettingsProps> = ({ taterMode, onTaterModeChange, onIdentityChange, origin, mode = 'plan', onUIPreferencesChange, externalOpen, onExternalClose, aiProviders = [], gitUser }) => {
   const [showDialog, setShowDialog] = useState(false);
